% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima}
\alias{arima}
\alias{predict}
\title{Using arima model to predict water supply time series in Shenzhen}
\usage{
predict(i)
}
\arguments{
\item{i}{a number between 2 and 21, means the selected column}
}
\value{
full dataset : hourly
}
\description{
Using arima model to predict water supply time series in Shenzhen
}
\examples{
\dontrun{
predict(2)
}
}
