% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dates}
\alias{dates}
\title{Compute how many days in month M, year Y}
\usage{
dates(Y, M)
}
\arguments{
\item{Y}{year}

\item{M}{month}
}
\value{
Days of Month M in Year Y
}
\description{
Compute how many days in month M, year Y, as a subfunction for water supply prediction
}
\examples{
\dontrun{
dates(2022,12)
}
}
