% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouplasso.R
\name{g_lasso}
\alias{g_lasso}
\title{group lasso function for variable selection}
\usage{
g_lasso(X, Y, lambda, p, J, index)
}
\arguments{
\item{X}{the design matrix with factor form}

\item{Y}{the response variable vector}

\item{lambda}{tuning parameter for penalty in group lasso}

\item{p}{numeric vector for factor sizes}

\item{J}{number of factors}

\item{index}{index for element position of factors, computed from p}
}
\value{
estimate of parameter beta in group lasso model
}
\description{
group lasso function for variable selection
}
