\name{StatComp22036-package}
\alias{StatComp22036-package}
\alias{StatComp22036}
\docType{package}
\title{
  Three functions for StatComp by 22036
}
\description{
  Three R and Rcpp function are included in this package, for goup lasso variables selection, water supply time series prediction and handwritten merge sort.
}
\details{
  Three R and Rcpp function are included in this package, for goup lasso variables selection, water supply time series prediction and handwritten merge sort.
}
\author{
22036

Maintainer: 22036 <1834326801@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
