% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima}
\alias{arima}
\alias{predict}
\title{Using complex arima model to predict water supply time series in Shenzhen.}
\usage{
predict(i)
}
\arguments{
\item{i}{a number between 2 and 21, means the selected neighborhood to train and predict, which is a column in the dataset.}
}
\value{
full dataset: hourly, including training data and predicted result,
}
\description{
Using complex arima model to predict water supply time series in Shenzhen.
}
\examples{
\dontrun{
predict(2)
}
}
