% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouplasso.R
\name{g_lasso}
\alias{g_lasso}
\title{Group lasso function for variable selection.}
\usage{
g_lasso(X, Y, lambda, p, J, index)
}
\arguments{
\item{X}{the design matrix with factor form.}

\item{Y}{the response variable vector.}

\item{lambda}{the tuning parameter for penalty in group lasso regression.}

\item{p}{the numeric vector for factor size.}

\item{J}{the number of factors.}

\item{index}{the index for element position of factors, computed from factor size p.}
}
\value{
the estimate of parameter beta in group lasso model
}
\description{
The Group lasso function for variable selection considering factors.
}
